1: Added values 3-5 as variables, and added 3-5 in arithmetic line.
2: Turned number variable into an integer.

3:
Error 1: first number variable isn't an integer
Error 2: second number variable is missing a
parenthesis for the integer function.
Error 3: missing a set of parentheses for the print function in line 13.
Error 4: number2 variable is capitalized wrongly in line 13.
Error 5: no quotation marks in line 14 for the print function.
Error 6: multiplication is formatted incorrectly, not using a " * ".

4: Added base and height variables using inputs converted to integers.
Then I did arithmetic in a print fucntion.

5: Added first and last name variables using an input
converted to a string.
Then I made a sentence combining the two variables in a print function.